flask_api/requirements.txt
gunicorn
Flask
pymongo
bcrypt
pyjwt[crypto]
Authlib
requests
mailjet_rest
itsdangerous